backend/requirements.txt
django>=4.0.6<4.1
djangorestframework>=3.13<3.14
python-dotenv>=0.20.0<0.21.0